Add files via upload
This script will go through all of your contacts and email the contact of your choice. This script can easily send emails to all contacts in your contacts folder and if you add "windows task scheduler" to schedule this program, emails can be sent automatically on a daily basis if you please.